backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Neo4j
neo4j==5.15.0
py2neo==2021.2.4

# Vector Store & Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2
chromadb==0.4.18

# Web Scraping
scrapy==2.11.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0
requests==2.31.0

# Document Processing
PyMuPDF==1.23.8
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0

# NLP & ML
spacy==3.7.2
transformers==4.36.0
torch==2.1.1
numpy==1.25.2
pandas==2.1.4

# LLM Integration
google-generativeai==0.3.2
openai==1.3.0
langchain==0.0.350
langchain-community==0.0.10

# Task Queue
celery==5.3.4
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
click==8.1.7
tqdm==4.66.1
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0
